feat : adding mvc architecture to the app
