+++ package.json
@@ -27,6 +27,7 @@
   },
   "dependencies": {
     "core-js": "^3.22.8",
+    "fractional": "^1.0.0",
     "regenerator-runtime": "^0.13.9"
   }
 }
